feat: All SonarQube issues were fixed

+++ src/main/java/com/pteriscope/webservice/entities/patient/service/PatientServiceImpl.java
@@ -1,11 +1,12 @@
 package com.pteriscope.webservice.entities.patient.service;
 
 import com.pteriscope.webservice.entities.patient.domain.persistence.PatientRepository;
-import com.pteriscope.webservice.exception.CustomException;
+import com.pteriscope.webservice.exception.PsRequestException;
 import com.pteriscope.webservice.entities.patient.domain.model.entity.Patient;
 import com.pteriscope.webservice.entities.patient.domain.serivces.PatientService;
 import com.pteriscope.webservice.entities.specialist.domain.model.entity.Specialist;
 import com.pteriscope.webservice.entities.specialist.domain.persistence.SpecialistRepository;
+import com.pteriscope.webservice.util.PsConstants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -16,46 +17,50 @@ import java.util.Optional;
 @Service
 public class PatientServiceImpl implements PatientService {
 
+    private final PatientRepository patientRepository;
+    private final SpecialistRepository specialistRepository;
+
     @Autowired
-    private PatientRepository patientRepository;
-    @Autowired
-    private SpecialistRepository specialistRepository;
+    public PatientServiceImpl(PatientRepository patientRepository, SpecialistRepository specialistRepository) {
+        this.patientRepository = patientRepository;
+        this.specialistRepository = specialistRepository;
+    }
 
     @Override
     public Patient createPatient(Long specialistId, Patient patient) {
         Optional<Specialist> specialist = specialistRepository.findById(specialistId);
         if (specialist.isPresent()) {
             if(patientRepository.existsByDni(patient.getDni()))
-                throw new CustomException(HttpStatus.BAD_REQUEST, "Ya existe un paciente con ese DNI");
+                throw new PsRequestException(HttpStatus.BAD_REQUEST, "Ya existe un paciente con ese DNI");
 
             patient.setSpecialist(specialist.get());
             return patientRepository.save(patient);
         }
         else{
-            throw new CustomException(HttpStatus.BAD_REQUEST, "Specialist with ID " + specialistId + " does not exist");
+            throw new PsRequestException(HttpStatus.BAD_REQUEST, "Specialist with ID " + specialistId + " does not exist");
         }
     }
 
     @Override
     public Patient getPatient(Long patientId) {
         return patientRepository.findById(patientId)
-                .orElseThrow(() -> new CustomException(HttpStatus.BAD_REQUEST, "Patient not found"));
+                .orElseThrow(() -> new PsRequestException(HttpStatus.BAD_REQUEST, PsConstants.PATIENT_NOT_FOUND));
     }
 
     @Override
     public List<Patient> getPatientFromSpecialist(Long specialistId) {
         Specialist specialist = specialistRepository.findById(specialistId)
-                .orElseThrow(() -> new CustomException(HttpStatus.BAD_REQUEST, "Specialist not found"));
+                .orElseThrow(() -> new PsRequestException(HttpStatus.BAD_REQUEST, PsConstants.SPECIALIST_NOT_FOUND));
         return patientRepository.getPatientsBySpecialist(specialist);
     }
 
     @Override
     public Patient updatePatient(Long patientId, Patient updatedPatient) {
         Patient existingPatient = patientRepository.findById(patientId)
-                .orElseThrow(() -> new CustomException(HttpStatus.BAD_REQUEST, "Patient not found"));
+                .orElseThrow(() -> new PsRequestException(HttpStatus.BAD_REQUEST, PsConstants.PATIENT_NOT_FOUND));
 
         if(patientRepository.existsByDni(updatedPatient.getDni()) && !updatedPatient.getDni().equals(existingPatient.getDni()))
-            throw new CustomException(HttpStatus.BAD_REQUEST, "Ya existe un paciente con ese DNI");
+            throw new PsRequestException(HttpStatus.BAD_REQUEST, "Ya existe un paciente con ese DNI");
 
         existingPatient.setFirstName(updatedPatient.getFirstName());
         existingPatient.setLastName(updatedPatient.getLastName());
@@ -68,8 +73,9 @@ public class PatientServiceImpl implements PatientService {
 
     @Override
     public void deletePatient(Long patientId) {
-        patientRepository.findById(patientId)
-                .orElseThrow(() -> new CustomException(HttpStatus.BAD_REQUEST, "Patient not found"));
+        if (!patientRepository.existsById(patientId)) {
+            throw new PsRequestException(HttpStatus.BAD_REQUEST, PsConstants.PATIENT_NOT_FOUND);
+        }
         patientRepository.deleteById(patientId);
     }
 }
